Cleanup IOleControl interface
